появляется ошибка и проект не собирается....не знаю что натворила...посмотрите пожалуйста код..а то что-то уже не соображаю
