+++ app/src/main/res/layout/activity_main.xml
@@ -9,31 +9,17 @@
     android:padding="@dimen/default_margin"
     tools:context=".MainActivity">
 
-    <RadioGroup
-        android:id="@+id/radioButtons"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/default_margin"
-        android:background="@drawable/rectangle_button"
-        android:gravity="end"
-        android:padding="@dimen/default_margin"
-        android:textColor="@color/black"
-        android:textSize="@dimen/textSize">
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="111dp"
+        android:orientation="horizontal">
 
-        <com.google.android.material.radiobutton.MaterialRadioButton
-            android:id="@+id/radioButtonMaterialLight"
+        <com.google.android.material.textview.MaterialTextView
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/material_light" />
-
-        <com.google.android.material.radiobutton.MaterialRadioButton
-            android:id="@+id/radioButtonMaterialDark"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/material_dark" />
-
-    </RadioGroup>
-
+            android:layout_height="95dp"
+            android:background="@drawable/rectangle_button"
+            android:padding="@dimen/default_margin" />
+    </LinearLayout>
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -43,7 +29,7 @@
         <com.google.android.material.textview.MaterialTextView
             android:id="@+id/text"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="95dp"
             android:background="@drawable/rectangle_button"
             android:gravity="end"
             android:hint="@string/hint_0"
@@ -177,8 +163,10 @@
             android:orientation="horizontal">
 
             <com.google.android.material.button.MaterialButton
+                android:id="@+id/button_choice_of_theme"
                 style="@style/my_style_calc_button_numeral"
-                android:visibility="invisible" />
+                android:text="@string/theme"
+                android:textSize="@dimen/textSizeTheme" />
 
             <com.google.android.material.button.MaterialButton
                 android:id="@+id/button_zero"
