+++ app/src/main/res/values-night/strings.xml
@@ -1,3 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">Calculator</string>
     <string name="Settlement_history">Здесь должна сохраняться история расчетов</string>
