+++ app/src/main/res/values/dimens.xml
@@ -3,6 +3,7 @@
     <dimen name="half_margin">8dp</dimen>
     <dimen name="default_margin">16dp</dimen>
     <dimen name="double_margin">32dp</dimen>
+    <dimen name="textSizeTheme">15 sp</dimen>
     <dimen name="textSize">24sp</dimen>
     <dimen name="size_text_view">30sp</dimen>
