+++ app/src/main/res/values/styles.xml
@@ -10,8 +10,6 @@
         <item name="android:textColor">@color/button_numeral</item>
         <item name="backgroundTint">#2500BCD4</item>
         <item name="cornerRadius">10dp</item>
-
-
     </style>
 
     <style name="my_style_calc_button_sign" parent="my_style_calc_button_numeral">
